requirements.txt
pymysql
sqlalchemy
fastapi
uvicorn
python-dotenv